Minor bug fix learning material

+++ epl343/BookingSystem/views.py
@@ -196,7 +196,7 @@ def dashboard(request):
 
 
 def addLearningMaterial(request):
-   # if request.user.is_authenticated and request.user.is_superuser::
+    if request.user.is_authenticated and request.user.is_staff:
         if request.method == "GET":#epistrefis forma
             return render(request, "addlearningmaterial.html", {'materialform': LearningMaterialForm()})
         if request.method == "POST":
@@ -207,45 +207,51 @@ def addLearningMaterial(request):
             else:#alios epestrepse error
                 message="error"
             return render(request, "addlearningmaterial.html", {'materialform': LearningMaterialForm(),'message':message})
-
 def getAllLearningMaterial(request):
-   # if request.user.is_authenticated and request.user.is_superuser::
-    allmaterial=LearningMaterial.objects.all()
-    return render(request, "getalllearningmaterial.html", {'allmaterials': allmaterial})
+    if request.user.is_authenticated and request.user.is_staff:
+        allmaterial=LearningMaterial.objects.all()
+        return render(request, "getalllearningmaterial.html", {'allmaterials': allmaterial})
 def addUserToLearningMaterial(request,id):
-   # if request.user.is_authenticated and request.user.is_superuser::
-    if request.method == "GET":#epistrefis forma
-        usersall=MyUser.objects.all()#epestrepse olus tus xristes
-        return render(request, "addusertomaterial.html", {'usersall': usersall})
-    if request.method == "POST":#epistrefis forma
-        usersall=MyUser.objects.all()
-        material=LearningMaterial.objects.filter(pk=id)
-        LearningMaterialReference.objects.all().filter(LearningMaterial__id=id).delete()
-        for x in usersall:
-            if(str(x.id) in request.POST):
-                LearningMaterialReference(User=x,LearningMaterial=material[0]).save()
-        return render(request, "addusertomaterial.html", {'usersall': usersall,'FilesLearningMaterialForm':FilesLearningMaterialForm()})
+    if request.user.is_authenticated and request.user.is_staff:
+        if request.method == "GET":#epistrefis forma
+            usersall=MyUser.objects.all()#epestrepse olus tus xristes
+            return render(request, "addusertomaterial.html", {'usersall': usersall})
+        if request.method == "POST":#epistrefis forma
+            usersall=MyUser.objects.all()
+            material=LearningMaterial.objects.filter(pk=id)
+            LearningMaterialReference.objects.all().filter(LearningMaterial__id=id).delete()
+            for x in usersall:
+                if(str(x.id) in request.POST):
+                    LearningMaterialReference(User=x,LearningMaterial=material[0]).save()
+            return render(request, "addusertomaterial.html", {'usersall': usersall,'FilesLearningMaterialForm':FilesLearningMaterialForm()})
 def addFileToMaterial(request):
-   # if request.user.is_authenticated and request.user.is_superuser::
-    if request.method == "GET":#epistrefis forma
-        return render(request, "addfiletomaterial.html", {'FilesLearningMaterialForm':FilesLearningMaterialForm()})
-    if request.method == "POST":#epistrefis forma
-        form=FilesLearningMaterialForm(request.POST,request.FILES)
-        msg=None
-        if form.is_valid():
-            msg="succeed"
-            form.save()
-        else:
-            msg="failed"
-        return render(request, "addfiletomaterial.html", {'FilesLearningMaterialForm':FilesLearningMaterialForm(),'message':msg})
+    if request.user.is_authenticated and request.user.is_staff:
+        if request.method == "GET":#epistrefis forma
+            return render(request, "addfiletomaterial.html", {'FilesLearningMaterialForm':FilesLearningMaterialForm()})
+        if request.method == "POST":#epistrefis forma
+            form=FilesLearningMaterialForm(request.POST,request.FILES)
+            msg=None
+            if form.is_valid():
+                msg="succeed"
+                form.save()
+            else:
+                msg="failed"
+            return render(request, "addfiletomaterial.html", {'FilesLearningMaterialForm':FilesLearningMaterialForm(),'message':msg})
 def userViewMaterial(request):
-    #if request.user.is_authenticated
-    learningmat=LearningMaterialReference.objects.filter(User__id=request.user.id)
-    return render(request, "viewlearningmaterial.html", {'learningmat':learningmat})
+    if request.user.is_authenticated:
+        learningmat=LearningMaterialReference.objects.filter(User__id=request.user.id)
+        return render(request, "viewlearningmaterial.html", {'learningmat':learningmat})
 def viewmaterial(request,id):
-    learningmat=FilesLearningMaterial.objects.filter(LearningMaterialFK__id=id)
-    #na kano ena elegxo oti anikoun ston xristi
+    learningmat=FilesLearningMaterial.objects.filter(LearningMaterialFK__id=id)#vrisko to antikimeno
+    if(len(learningmat)!=0):
+        learningMaterialMainObject=learningmat[0].LearningMaterialFK#dixnw ston dixti tou learningmaterial antikimenou
+        references=LearningMaterialReference.objects.all().filter(LearningMaterial=learningMaterialMainObject).filter(User=request.user)#vrisko an iparxi erotima gia afton ton xristi
+        if(len(references)==0):#an to material den iparxi ston xristi den tha tou emfaniso kati
+            learningmat=None
+    else:#den iparxun arxia
+        message="no files"
     return render(request, "viewmaterialfiles.html", {'learningmat':learningmat})
+
